種々の修正
